Update modules to the go 1.16.3 code
This is the output of running `script/extract` locally on my Mac which
is running `go 1.16.3`.

I suspect this pulls in a few fixes/changes in how go module machinery
works.

I was uncertain whether to use tip of master on `go`, but thought better
to use a stable version as that'd likely be more predictable.