Add historical explanation of the code
Explain the purpose of this code. 

This is in preparation for archiving this repo now that https://github.com/dependabot/dependabot-core/pull/4434/ has landed.

Because archived repos can't have issues/PR's filed on them, so once archived if the reason isn't on the readme it's very confusing
to drive-by committers.